+++ scripts/deploy/hardhat/01_deploy_llama_framework.ts
@@ -1,11 +1,6 @@
-// We require the Hardhat Runtime Environment explicitly here. This is optional
-// but useful for running the script in a standalone fashion through `node <script>`.
-//
-// When running the script with `npx hardhat run <script>` you'll find the Hardhat
-// Runtime Environment's members available in the global scope.
 import { ethers } from "hardhat";
 
-import { LlamaFrameworkAddress } from "../../types/Llama";
+import { LlamaFrameworkAddress } from "../../types/deploy";
 
 async function main(): Promise<{ llamaFramework: LlamaFrameworkAddress }> {
