+++ scripts/deploy/hardhat/02_deploy_llama_instance.ts
@@ -1,11 +1,4 @@
-// We require the Hardhat Runtime Environment explicitly here. This is optional
-// but useful for running the script in a standalone fashion through `node <script>`.
-//
-// When running the script with `npx hardhat run <script>` you'll find the Hardhat
-// Runtime Environment's members available in the global scope.
 import { ethers } from "hardhat";
-import { defaultAbiCoder, AbiCoder } from "@ethersproject/abi";
-
 import { encodeRelativeStratigyConfig, encodeAccountConfig, encodeBytes32 } from "../../_utils/llama";
 import LLAMA from "../../../config/llama.json";
