+++ scripts/deploy/hardhat/04_deploy_vesting_vault.ts
@@ -1,10 +1,4 @@
-// We require the Hardhat Runtime Environment explicitly here. This is optional
-// but useful for running the script in a standalone fashion through `node <script>`.
-//
-// When running the script with `npx hardhat run <script>` you'll find the Hardhat
-// Runtime Environment's members available in the global scope.
 import { ethers } from "hardhat";
-
 import { DeployLocal } from "../../types/deploy";
 
 async function main({ deployed }: { deployed: DeployLocal }) {
