+++ scripts/deploy/hardhat/05_deploy_oracle.ts
@@ -1,10 +1,4 @@
-// We require the Hardhat Runtime Environment explicitly here. This is optional
-// but useful for running the script in a standalone fashion through `node <script>`.
-//
-// When running the script with `npx hardhat run <script>` you'll find the Hardhat
-// Runtime Environment's members available in the global scope.
 import { ethers } from "hardhat";
-import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { DeployLocal } from "../../types/deploy";
 
 async function main({ deployed }: { deployed: DeployLocal }) {
@@ -18,8 +12,8 @@ async function main({ deployed }: { deployed: DeployLocal }) {
   const PriceOracleRouterFactory = await ethers.getContractFactory("PriceOracleRouter");
   const priceOracleRouter = await PriceOracleRouterFactory.deploy();
 
-  const CCNotePriceOracleFactory = await ethers.getContractFactory("CCNotePriceOracle");
-  const ccNotePriceOracle = await CCNotePriceOracleFactory.deploy(deployed.CNOTE!);
+  const VCNotePriceOracleFactory = await ethers.getContractFactory("VCNotePriceOracle");
+  const vcNotePriceOracle = await VCNotePriceOracleFactory.deploy(deployed.CNOTE!);
 
 
   ////////////////////////////////
@@ -30,7 +24,7 @@ async function main({ deployed }: { deployed: DeployLocal }) {
   return {
     oracle: {
       priceOracleRouter: priceOracleRouter.address,
-      ccNotePriceOracle: ccNotePriceOracle.address,
+      vcNotePriceOracle: vcNotePriceOracle.address,
     }
   }
 }
