+++ scripts/deploy/hardhat/07_deploy_interest_model.ts
@@ -1,10 +1,4 @@
-// We require the Hardhat Runtime Environment explicitly here. This is optional
-// but useful for running the script in a standalone fashion through `node <script>`.
-//
-// When running the script with `npx hardhat run <script>` you'll find the Hardhat
-// Runtime Environment's members available in the global scope.
 import { ethers } from "hardhat";
-import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { DeployLocal } from "../../types/deploy";
 import MODELS from "../../../config/interestModels.json";
